clean(BookingSideBar): flat structure and refactor keys
