+++ src/styles/colors.ts
@@ -1,4 +1,5 @@
 export const TEXT = "#3E3E3E";
+export const SECONDARY_TEXT = "#8B8B8B";
 
 export const WHITE = "#fff";
